requirements.txt
# Core dependencies for Luminate API Analyzer

# Web framework for creating the user interface
streamlit>=1.28.0

# HTTP library for API requests
requests>=2.31.0

# Data manipulation and analysis
pandas>=2.0.0

# Environment variable management
python-dotenv>=1.0.0

